infrabot-backend/pyproject.toml
[project]
name = "infrabot-agent"
version = "0.1.1"
description = "LLM-powered IT help desk service"
authors = [
    {name = "Sarath R", email = "sarath.r@prodapt.com"},

]
packages = [
    {include = "app", from = "."}
]

readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn>=0.34.2,<0.35.0",
    "openai>=1.82.1,<2.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "pydantic>=2.11.5,<3.0.0",
    "langchain>=0.3.25,<0.4.0",
    "langchain-community>=0.3.25,<0.4.0",
    "langchain-openai>=0.3.22,<0.4.0",
    "langchain-milvus (>=0.2.1,<0.3.0)",
    "python-dotenv>=1.0.0,<2.0.0",
    "unstructured[pdf]~=0.17.2",
    "unstructured[ocr] == 0.17.2",
    "pdf2image >=1.17.0,<2.0.0",
    "pdfminer-six==20250506",
    "unstructured-inference == 0.8.10",
    "pi-heif>=0.22.0,<0.23.0",
    "pymilvus>=2.5.7,<3.0",
    "pytesseract == 0.3.10",
    "unstructured-pytesseract (>=0.3.15,<0.4.0)",
    "setuptools<81"
]

[tool.poetry]
name = "infrabot-agent"
version = "0.1.1"
description = "LLM-powered IT help desk service"
authors = ["Sarath R <sarath.r@prodapt.com>"] 
readme = "README.md" # Mirroring from [project]
packages = [
    {include = "app", from = "."}
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
openai = ">=1.82.1,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
pydantic = ">=2.11.5,<3.0.0"
langchain = ">=0.3.25,<0.4.0"
langchain-community = ">=0.3.25,<0.4.0"
langchain-openai = ">=0.3.22,<0.4.0"
langchain-milvus = ">=0.2.1,<0.3.0"
setuptools = "<81"
pypdf = ">=4.0.0,<5.0.0"
pdfminer-six = "==20250506"
pi-heif = "^0.22.0"
pdf2image = ">=1.17.0,<2.0.0"
unstructured-inference = "^0.8.10"
unstructured = {version = "^0.17.2", extras = ["ocr"]}
pymilvus = "^2.5.7"
pytesseract = "^0.3.10"
torch = {version = "2.7.1+cpu", source = "pytorch_cpu"}
torchvision = {version = "0.22.1+cpu", source = "pytorch_cpu"}


[tool.poetry.scripts]
infrabot-agent = "app.main:app"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
httpx = "^0.28.1"
pytest = "^8.3.5"
pytest-asyncio = "^1.0.0"
black = "^25.1.0"
mypy = "^1.16.0"
types-ujson = "^5.10.0.20250326"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "UP"]
ignore = ["D203", "D212"]
exclude = ["tests/*", "migrations/*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
